requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langchain==0.3.1
langchain-google-genai==1.0.10
langchain-pinecone==0.1.3
pinecone-client==5.0.1
pypdf2==3.0.1
python-multipart==0.0.9
pydantic==2.8.2